fix: handle optional flags from Argo
